+++ app/src/main/java/at/aau/se2/chessify/AndroidGameUI/BoardView.java
@@ -10,7 +10,6 @@ import android.app.Dialog;
 import android.content.Context;
 import android.media.MediaPlayer;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.View;
 import android.view.Window;
 import android.view.WindowManager;
@@ -30,17 +29,17 @@ import java.util.List;
 
 import at.aau.se2.chessify.Game;
 import at.aau.se2.chessify.R;
-import at.aau.se2.chessify.chessLogic.board.ChessBoard;
-import at.aau.se2.chessify.chessLogic.board.Location;
-import at.aau.se2.chessify.chessLogic.board.Move;
-import at.aau.se2.chessify.chessLogic.pieces.Bishop;
-import at.aau.se2.chessify.chessLogic.pieces.ChessPiece;
-import at.aau.se2.chessify.chessLogic.pieces.King;
-import at.aau.se2.chessify.chessLogic.pieces.Knight;
-import at.aau.se2.chessify.chessLogic.pieces.Pawn;
-import at.aau.se2.chessify.chessLogic.pieces.PieceColour;
-import at.aau.se2.chessify.chessLogic.pieces.Queen;
-import at.aau.se2.chessify.chessLogic.pieces.Rook;
+import at.aau.se2.chessify.chess_logic.board.ChessBoard;
+import at.aau.se2.chessify.chess_logic.board.Location;
+import at.aau.se2.chessify.chess_logic.board.Move;
+import at.aau.se2.chessify.chess_logic.pieces.Bishop;
+import at.aau.se2.chessify.chess_logic.pieces.ChessPiece;
+import at.aau.se2.chessify.chess_logic.pieces.King;
+import at.aau.se2.chessify.chess_logic.pieces.Knight;
+import at.aau.se2.chessify.chess_logic.pieces.Pawn;
+import at.aau.se2.chessify.chess_logic.pieces.PieceColour;
+import at.aau.se2.chessify.chess_logic.pieces.Queen;
+import at.aau.se2.chessify.chess_logic.pieces.Rook;
 import at.aau.se2.chessify.network.LifeCycleObserver;
 import at.aau.se2.chessify.network.WebSocketClient;
 import at.aau.se2.chessify.network.dto.GameDataDTO;
