+++ app/src/main/java/at/aau/se2/chessify/chess_logic/board/Location.java
@@ -1,4 +1,4 @@
-package at.aau.se2.chessify.chessLogic.board;
+package at.aau.se2.chessify.chess_logic.board;
 
 import java.util.ArrayList;
