+++ app/src/main/java/at/aau/se2/chessify/chess_logic/pieces/ChessMoveInterface.java
@@ -1,9 +1,9 @@
-package at.aau.se2.chessify.chessLogic.pieces;
+package at.aau.se2.chessify.chess_logic.pieces;
 
 import java.util.ArrayList;
 
-import at.aau.se2.chessify.chessLogic.board.ChessBoard;
-import at.aau.se2.chessify.chessLogic.board.Location;
+import at.aau.se2.chessify.chess_logic.board.ChessBoard;
+import at.aau.se2.chessify.chess_logic.board.Location;
 
 public interface ChessMoveInterface {
