+++ app/src/test/java/at/aau/se2/chessify/chessLogicTests/pieces/QueenTest.java
@@ -3,8 +3,8 @@ package at.aau.se2.chessify.chessLogicTests.pieces;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
-import at.aau.se2.chessify.chessLogic.board.*;
-import at.aau.se2.chessify.chessLogic.pieces.*;
+import at.aau.se2.chessify.chess_logic.board.*;
+import at.aau.se2.chessify.chess_logic.pieces.*;
 
 import org.junit.Before;
 import org.junit.Test;
